RemindersFragment and SettingsFragment creation, NavBottomView migration to MainActivity, and some refactors in the whole views, extensions.kt added to common.
