+++ app/src/main/java/com/example/rememberme/reminders/RemindersFragment.kt
@@ -1,0 +1,8 @@
+package com.example.rememberme.reminders
+
+import androidx.fragment.app.Fragment
+import com.example.rememberme.R
+
+class RemindersFragment : Fragment(R.layout.fragment_reminders) {
+
+}
