+++ app/src/main/java/com/example/rememberme/settings/SettingsFragment.kt
@@ -1,0 +1,8 @@
+package com.example.rememberme.settings
+
+import androidx.fragment.app.Fragment
+import com.example.rememberme.R
+
+class SettingsFragment : Fragment(R.layout.fragment_settings) {
+
+}
