+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/on_boarding_dest"
-        android:name="com.example.rememberme.OnBoardingFragment"
+        android:name="com.example.rememberme.onBoarding.OnBoardingFragment"
         android:label="fragment_on_boarding"
         tools:layout="@layout/fragment_on_boarding" >
         <action
@@ -16,7 +16,17 @@
     </fragment>
     <fragment
         android:id="@+id/menu_dest"
-        android:name="com.example.rememberme.MenuFragment"
+        android:name="com.example.rememberme.menu.MenuFragment"
         android:label="fragment_menu"
         tools:layout="@layout/fragment_menu" />
+    <fragment
+        android:id="@+id/reminders_dest"
+        android:name="com.example.rememberme.reminders.RemindersFragment"
+        android:label="fragment_reminders"
+        tools:layout="@layout/fragment_reminders" />
+    <fragment
+        android:id="@+id/settings_dest"
+        android:name="com.example.rememberme.settings.SettingsFragment"
+        android:label="fragment_settings"
+        tools:layout="@layout/fragment_settings" />
 </navigation>
